requirements.txt
streamlit>=1.32.0
supabase>=2.3.0
python-dotenv>=1.0.0
plotly>=5.18.0
pandas>=2.1.0
